+++ packages/jates/Behavior.ts
@@ -1,6 +1,6 @@
 import { WsPoolProv } from "^jab-express";
 import { ClientMessage, ServerMessage } from "^jatec";
-import { safeAll } from "^yapu";
+import { safeAll, safeAllWait } from "^yapu";
 import { TestFrameworkProv } from "./internal";
 
 // prov
@@ -19,14 +19,35 @@ export type BehaviorDeps = {
  *
  */
 export class Behavior implements BehaviorProv {
+  private currentWork: Set<Promise<unknown>> = new Set();
+
   constructor(private deps: BehaviorDeps) {}
 
   /**
    *
    */
-  public onShutdown = () =>
-    safeAll(
-      [this.deps.testFramework.kill(), this.deps.wsPool.shutdown()],
-      this.deps.onError
-    ).then(() => {}); //just for typing
+  public setWorking = (prom: Promise<unknown>) => {
+    this.currentWork.add(prom);
+
+    prom.finally(() => {
+      this.currentWork.delete(prom);
+    });
+  };
+
+  /**
+   *
+   */
+  public onShutdown = async () => {
+    //first kill framework to stop its work
+
+    await this.deps.testFramework.kill();
+
+    //wait for promises started by director
+
+    await safeAllWait(Array.from(this.currentWork), this.deps.onError);
+
+    //shutdown client connections
+
+    await this.deps.wsPool.shutdown();
+  };
 }