+++ packages/tests/all/jab-node NodeWs err 2.ja.ts
@@ -1,11 +1,9 @@
 import { TestProvision } from "^jarun";
 
-import { getNodeWs, getServer_test_app } from "../_fixture";
+import { getNodeWs } from "../_fixture";
 
-// doesn't close nodeWs or server
-export default (prov: TestProvision) =>
-  getServer_test_app(prov).then(() =>
-    getNodeWs(prov).then(() => {
-      // to avoid return
-    })
-  );
+// host invalid
+
+export default (prov: TestProvision) => {
+  getNodeWs(prov, "", { host: "" });
+};