+++ packages/tests/all/jab-node NodeWs err 4.ja.ts
@@ -1,9 +1,0 @@
-import { TestProvision } from "^jarun";
-
-import { getNodeWs } from "../_fixture";
-
-// host invalid
-
-export default (prov: TestProvision) => {
-  getNodeWs(prov, "", { host: "" });
-};